backend/requirements.txt
fastapi
uvicorn[standard]
clickhouse-connect
kafka-python
